# backend/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Items Service API
  version: 1.0.0
  description: REST API for Items, Categories, CSV import/export. Spring-friendly (Pageable, ETags, RFC-7807).
servers:
  - url: /api/v1
...
paths:
...
  /categories:
    get:
      tags: [Categories]
      summary: List/search categories (pageable)
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/sortCategory'
        - name: q
          in: query
          required: false
          description: Case-insensitive substring on name
          schema: { type: string }
      responses:
        '200':
          description: Paged categories
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CategoryPage' }
    post:
      tags: [Categories]
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCategoryRequest' }
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '400': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /categories/{id}:
    get:
      tags: [Categories]
      summary: Get category
      parameters:
        - $ref: '#/components/parameters/categoryPathId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '404': { $ref: '#/components/responses/Problem' }
    patch:
      tags: [Categories]
      summary: Rename category
      parameters:
        - $ref: '#/components/parameters/categoryPathId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RenameCategoryRequest' }
      responses:
        '200':
          description: Renamed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
    delete:
      tags: [Categories]
      summary: Delete category
      parameters:
        - $ref: '#/components/parameters/categoryPathId'
        - name: force
          in: query
          required: false
          schema: { type: boolean, default: false }
          description: Force delete even if in use (implementation-defined)
      responses:
        '204':
          description: No Content
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    categoryPathId:
      name: id
      in: path
      required: true
      description: Category id (UUID)
      schema: { type: string, format: uuid }
...
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 8 }
...
    sortCategory:
      name: sort
      in: query
      description: Sort by name,asc|desc
      schema: { type: string, example: name,asc }
...
  responses:
    Problem:
      description: Problem Details (RFC-7807)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  schemas:
...
    Category:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 100 }
    CategoryPage:
      type: object
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/Category' }
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1 }
        totalElements: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 1 }
        sort:
          type: array
          items: { type: string, example: name,asc }
    CreateCategoryRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
    RenameCategoryRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }

    # --- Problem / Import result ---
    ProblemDetail:
      type: object
      description: RFC-7807 Problem Details
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: object
          additionalProperties:
            type: string
...
security:
  - bearerAuth: []